requirements.txt
nats-python==0.4.0
kafka-python==2.0.1
jsonpickle==1.4.1
redis==3.5.3
protobuf==3.0.0
requests==2.24.0